set use selection to false as default
